Fixing code coverage issues. Adding forcaster strategies.

+++ day_trade/src/strategies/hold/bollinger_bands.rs
@@ -401,7 +401,7 @@ mod tests {
                     high: price + 0.2,
                     low: price - 0.2,
                     close: price,
-                    volume: 1000 + (minute % 10) * 100,
+                    volume: (1000 + (minute % 10) * 100) as u64,
                 },
             });
         }
